requirements.txt
# --- Data extraction ---
python-docx==0.8.11       # Đọc file .docx, cần thiết cho ingestion CV Word
PyPDF2==3.0.1             # Đọc file PDF, hỗ trợ CV PDF nếu có

# --- Text processing ---
regex                     # Biểu thức chính quy nâng cao, xử lý text
nltk                      # Tokenization, stopwords, xử lý cơ bản NLP
spacy                     # NLP, Named Entity Recognition, trích xuất entities (skills, roles)
textblob                  # Optional, hỗ trợ sentiment hoặc NLP đơn giản

# --- Machine Learning & Vectorization ---
scikit-learn              # TF-IDF vectorization, cosine similarity, scoring & ranking
numpy                     # Tính toán số học, vector hóa
pandas                    # Quản lý data frames, lưu trữ CV sau khi preprocess

# --- Deep Learning / Embeddings ---
sentence-transformers     # BERT/SBERT embeddings cho semantic similarity giữa CV và Job Description
faiss-cpu==1.12.0         # Index nhanh cho embeddings, tăng tốc tìm kiếm tương đồng semantic

# --- Visualization ---
matplotlib                # Vẽ biểu đồ, thống kê skill distribution, visual debug

# --- API / Frontend ---
Flask                     # Backend API để expose pipeline, có thể kết nối frontend
streamlit                 # Dashboard frontend, upload Job Description & hiển thị top CVs

# --- Utilities ---
python-dotenv             # Load config từ file .env, ví dụ threshold, top-K
tqdm                      # Thanh tiến trình khi xử lý nhiều file CV (>200)
